Deduplicate `struct Dav1dTaskContext_scratch_levels_pal_ac_interintra_edge` to `src/internal.rs`.
